Make the OLPC XO-1.75 boot script a bit more forgiving
This makes the script slighlty less picky about the compatible strings.

In general, if the root compatible strings contains ids that are known
supported by mainline/Debian kernel, we know the machine is going to boot
and we don't need to do further checks.

This makes the logic a bit simpler, but also ensures XO-4 (and
potentially other MMP boards such as Dell Ariel) will boot fine if it
provides a good device tree.