Add support for cross-building
Here is some minimal support to enable to build debian-installer on
another arch than the target (ppc64el on amd64 here).
I have been asked this enablement to do CI on the installer and its
components without having for instance a ppc64el machine and still be
able to generate the debian-installer especially on salsa.
Which works now, but other combinations may need specific changes or be
not possible at all.

For ppc64el generation on amd64, one now just need a :
sbuild --on amd64 --host ppc64el --profiles=cross,nocheck

The mini.iso just worked fine. It also worked for powerpc on amd64 but
I didn't test the result.