gen-hd-image: Add support for building GPT disk images
The gen-hd-image utility can be used to generate disk images that have a
special bootloader-related partition next to a normal data partition,
but it only supports MBR.

Add GPT support to gen-hd-image so it can be used to generate disk
images for chromebooks. Sfdisk's short form for the corresponding
--label argument is -X, use that here as well. Move the partition ID
code to when we definitely know the label and filesystem type. GPT also
uses the end of the disk for some data, so fixup filesystem size a bit
and write the filesystem into proper offsets in the disk image file.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>