gen-hd-image: Mark ChromeOS kernel partitions as bootable
A previous commit adds GPT support to gen-hd-image which would helps us
create more functional disk images for chromebooks. But we still need to
get an uncompressed image from gen-hd-image, mark it as bootable for
depthcharge, then compress it afterwards.

Marking the partition bootable is actually simple, make gen-hd-image do
it so we can delegate compressing the disk image to it as well.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>