Minimize kernel partition size in depthcharge images
The depthcharge images have so far had their ChromeOS kernel partition
size set to 512MiB because that's the size limit for the most recent
boards, in case people would want to reuse them for the installed
system. It's usually 150 MiB or so, 512MiB is too wasteful anyway.

That no longer makes that much sense to me. Furthermore, we have a nice
use case for a new second partition, but it needs to be bigger for that.
Make the kernel partition as small as possible, by using the size of the
depthcharge kernel image we will to write to it.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>