Build amd64/arm64 cdrom images for ChromeOS devices
The cdrom and cdrom/gtk builds can technically work as depthcharge
images, but were not added so far since it was unclear how. In the worst
case we could use two disks: write the depthcharge disk image to one, a
Debian installer CD/DVD image to the other, and boot with both plugged
in. However that is cumbersome.

Now that we have a second partition to use in the depthcharge disk
images, we can ask users to write or extract the netinst CD image into
it, which will make it possible use a single disk to boot the installer.

Add depthcharge images as build targets in the existing flavours that
build cdrom and graphical cdrom images. It seems more proper to copy the
existing flavour configs and package-lists into new "depthcharge" and
"depthcharge-gtk" flavours for the cdrom build, but that introduces new
make targets and this one is more consistent with how the netboot builds
handle depthcharge and concatenable images.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>