Support building board-specific depthcharge images
The generic depthcharge disk images can only work for recent devices
where a hard-coded size limit in their firmware is upgraded to 512MiB.
We want to support earlier arm64 devices that have an 32MiB limit.

It's possible to make the non-graphical arm64 cdrom files fit into a
32MiB limit by compressing the initramfs as xz and including less
device-tree files, with the trade off of making the images specific to
certain board families. In the future, we might be able to use U-Boot as
an intermediary bootloader for these devices which would also need
per-board builds, so this is not that much of a disadvantage.

Adjust the depthcharge image recipes to build for a list of boards from
a make variable. For now, only build the generic images for both
architectures.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>