Build arm64 cdrom depthcharge image for "gru" boards
Hardware support for "gru" family of arm64 chromebooks is mostly
complete in Debian as I have been working on that, but it has a 32MiB
size limit for its images. The non-graphical cdrom image can fit into
it, but barely. Enable building the depthcharge image for "gru"
chromebooks for this flavour.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>