Build very small hd-media images as a 'tiny' flavour
By removing more modules from the hd-media image we can get it small
enough to fit into the size limit for earlier ARM64 Chromebooks
supported in Debian. If we remove more things that we really shouldn't,
we can get it to work with an even smaller size limit for the earliest
AMD64 Chromebooks.

Add a new "tiny" hd-media flavour that does removes a lot of modules
from the installer and uses xz to compress the initramfs. Build it for
"Gru" and "Hana" Chromebooks on ARM64, and build one generic image for
AMD64 because specifying a board doesn't affect image size there.

A better solution would be to create a very small initramfs which does
nothing except find and chainload the actual initramfs we want, e.g. as
in postmarketOS.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>