Build arm64/amd64 hd-media images for ChromeOS devices
The hd-media and hd-media/gtk builds are slightly easier to work with
than the cdrom versions, since we can just copy a Debian installer iso
file into the second partition instead of extracting or dd'ing the file.

Add depthcharge images as build targets in the existing flavours that
build hd-media and graphical hd-media images.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>