Make x86 depthcharge images configure partman to use GPT partitioning
The ChromeOS bootloader's verified boot mechanism can only boot from GPT
partitions, so installing a system that boots with that must partition
its disks using GPT. However non-EFI x86 systems default to MBR and
the question that sets it is asked with a low priority.

Pass a kernel command line argument to set the default disk label to GPT
in partman. However, a better solution would be to define ChromeOS
devices as a new subarch in archdetect and then check that subarch in
partman-partitioning.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>