Create a second partition in depthcharge images
The netboot depthcharge images would benefit from a second partition
that we can put firmware onto. It could be created after writing our
images to disk, but with some tools adding partitions ends up wiping the
GUID-specific GPT partition attributes of the kernel partition, making
it unbootable.

Resize the depthcharge disk images to close to 1G like in gen-hd-image,
and create a FAT32 Linux data partition filling in the free space. The
size calculations are a bit complex because GPT reserves 33 sectors at
the end of the disk and sfdisk aligns partitions to 1MiB.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>