Fix missing unicode font for unsigned grub
In commit a4dc8c0fe7ad1a0c1506125ad9985f78819a1bb2, the default font was
changed to unicode. It worked for signed grub efi images, which is a
"monolithic" grub that embeds everything, including the unicode font.

However for derivatives that don't use the signed grub from Debian, what
we get is a "traditional" grub (ie. not monolithic), so the unicode font
must be installed at the right location. And that's not the case, all we
have is grub/font.pf2, while what we need is grub/fonts/unicode.pf2.

This commit fixes it. However, as a side-effect, it also breaks
debian-cd, which expects to find file(s) grub/*.pf2, and fails if
there's no such:

    build/debian-cd stdout:   Adding EFI boot code for amd64 on CD1
    build/debian-cd stderr: cp: cannot stat 'boot1/isolinux/grub-amd64/*.pf2': No such file or directory
    build/debian-cd stdout:       FAILED: error 1
    build/debian-cd stderr:     Failed to start disc 1, error 256
    build/debian-cd stderr: make: *** [Makefile:487: image-trees] Error 1

Therefore a fix in debian-cd must also be applied.

Closes: #1038862