Add md-modules to arm64 cdrom builds
U-Boot is adding support for HTTP(s) boot. The installation media is made
available as a /dev/pmem device: in order for the installer to be able to mount
it, we need to have pmem modules in the installer initrd.

This is one of the MRs needed to support U-Boot HTTP(s) boot, the other two
being:

- [arm64] Add {of,nd}_pmem to arm64 md-modules udeb
  https://salsa.debian.org/kernel-team/linux/-/merge_requests/1167

- Add /dev/pmem devices to list-devices-linux in debian-installer-utils (TBD)

See https://www.linaro.org/blog/ledge-blogs-uefi-http-and-https-boot-in-u-boot/