Add debian-installer-images tarball to CI reprotest artifacts
The salsa-ci pipeline runs reprotest without an artifact pattern by
default, and reprotest defaults to testing "*.deb" generated as a result
of the build. A recent commit to the pipeline adds a new variable we can
use to specify the artifact pattern argument to be passed to reprotest.

The debian-installer build generates a tarball of images, so set the
variable to include it. For unknown reasons diffoscope exits with -9,
so we can't enable it yet to get a description of the difference other
than "Binary files differ".

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
(cherry picked from commit 2eeff1d84ef33570ffc58f58b983d3ef7af8fd74)
Signed-off-by: Holger Levsen <holger@layer-acht.org>