hppa: Do not use sub-architectures
This reverts most of the previous commit d9fe68180f42 ("hppa: Rewrite
configuration to support sub-architectures") which intended to switch
hppa to use sub-architectures. Main reasons to revert are:
a) It's not useful and wastes space. We basically store the same initrd
   twice for each of the sub-architectures, and
b) the hppa/parisc bootloader (palo) is not prepared to handle two
   initrds when booting.