salsa-CI: avoid use of (before|after)_script
https://salsa.debian.org/salsa-ci-team/pipeline/-/merge_requests/566
introduces the use of after_script:, which clashes with our usage of it.

They are using after_script: in order to let them have code that is run
even if the main script: fails, whereas we only want to execute our code
at the end of a successful run, so this !reference approach in script:
is actually better for our needs anyway.

Also apply similar changes to avoid use of before_script, as doing so
now avoids the risk that we hit a similar issue if they were to start
using before_script in the upstream pipeline code.