Ensure destination directory exists before writing udeb lists
When an architecture has no images to build, the BASE_DEST directory
doesn't get created by the time we try to write the udeb list and update
the checksum files. Make sure the directory exists before writing files
into it.

Alternatively, we could skip creating these files for empty builds and
create an empty tarball instead, or even skip creating an empty tarball.
But I don't know if the daily-build infrastructure can handle that. And
having the checksum files is a confirmation that a build was successful,
albeit with nothing to build.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>