Improve support for speech synthesis.
Intel is weird, and some sound modules might require i915 bits to work
properly. Until this is investigated more thoroughly (possibly adjusting
sound-modules contents and/or dependencies), let's include some modules
manually. Do that conditionally, for graphical installer builds (that's
where speech synthesis support is deployed), when the i915 module is
confirmed to be available (i.e. x86 only).

With thanks to Cleverson Casarin Uliana for the initial report, to
Samuel Thibault for looping us in, and to Roland Clobus for sorting
out which modules are needed to get this up and running.

Link: https://lists.debian.org/debian-boot/2025/04/msg00357.html