Use a block size of 1MiB for xz with default compression preset
According to the xz manual page, the only threading method it supports
is to split the input into blocks and compress them independently. The
size of these blocks is controlled by the --block-size= option, which
defaults to the higher of 1MiB or three times the dictionary size, so
about 192MiB for the -9 compression preset. Since all initrd files we
create are less than that, it happens to always run a single thread.

A naive attempt at the 'optimal' block size would be choosing the file
size divided by the system's CPU cores. But empirically, the difference
between using 1MiB and 16MiB as block size is about 2-3% of the
compressed size for our input files within 50-150MiB range, while the
16MiB runs can take about 2-4 times as long.

Set the xz block size to 1MiB, so we can have the speed up at the cost
of compressed file sizes. At that point using the -9 compression preset
results in the almost same size files as the default -6, so drop that
too. Where we truly need the file size to be small, we can override it.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>