Use fakeroot instead of Rules-Requires-Root: binary-targets
This makes the package not FTBFS with debrebuild and on
reproduce.debian.net.

debrebuild uses the buildinfo file generated on the buildd to reproduce
the package. But the buildinfo file does not safe the version of the
fakeroot package used so it can't use that and uses the unshare root
instead. unshare root can't handle mknod as done in build/Makefile so
the package FTBFS there. Instead this depends on fakeroot and rely on
build/Makefile re executing with fakeroot.

This reverts 680ab3ddc.

(cherry picked from commit 536db97fe06401e82be0ad9015f3c76866aae996)
Signed-off-by: Holger Levsen <holger@layer-acht.org>