Bootstrap p-u testing branch for trixie.
Cherry-pick the first commit of the kibi/bookworm-proposed-updates
branch (and amend it, see below). The other commits were just merges
from the bookworm branch.

Patch /usr/share/keyrings/debian-archive-keyring.pgp instead of its
debian-archive-keyring.gpg counterpart: the latter used to be a regular
file, but it's become a symlink to the former; modifying it turns the
symlink into a regular file which is disconnected from the former,
leading to failures to verify downloads.