artwork: Drop the version from the Ceratopsian and Homeworld boot screens
It's misleading for the d-i snapshot that will install testing/unstable
to identify itself as Debian 13. Until Debian 14 artwork becomes
available, delete the "13" from the Ceratopsian boot splash artwork,
and move the logotype across to compensate.

Similarly, remove the "11" from the Homeworld artwork for completeness.

This is analogous to commit 074ca9ca "Drop the “12” version from the
Emerald boot screen", but for forky.

Closes: #1118488
Signed-off-by: Simon McVittie <smcv@debian.org>