salsa-CI: restore miniISO's ability to access aptly repo
This is a change is prompted by the salsa-CI pipeline switching to
sbuild with the unshare backend, which no longer pays attention to edits
that are applied to the source just before running sbuild.

The APTLY_* settings needed to achieve this are passed into sbuild's env
by adding settings to sbuild's config file.

In debian/rules, conditional upon the 'pkg.debian-installer.miniiso'
profile being defined, this creates a sources.list.udeb.local file that
includes the current salsa-CI pipeline's aptly repo, and also ensures
that the miniISO being created includes a preseed.cfg that also points
to that aptly repo, so that the installed target system will be able to
make use of the repo as well.

this also adds a conditional dependancy on ca-certificates so that the
aptly repo (which is via an https URL) will be seen to have a valid
certificate (otherwise apt refuses to trust it).